Test app refactoring
