+++ examples/ongoingpostbox/src/main/java/me/digi/ongoingpostbox/data/localaccess/MainLocalDataAccess.kt
@@ -4,6 +4,9 @@ import io.reactivex.rxjava3.core.SingleTransformer
 import me.digi.sdk.entities.DMEOAuthToken
 import me.digi.sdk.entities.DMEPostbox
 
+/**
+ * Check [MainLocalDataAccessImpl] for more information
+ */
 interface MainLocalDataAccess {
     fun getCachedCredential(): DMEOAuthToken?
     fun getCachedPostbox(): DMEPostbox?
