--- examples/ongoingpostbox/src/main/java/me/digi/ongoingpostbox/features/base/BaseActivity.kt
@@ -1,5 +1,0 @@
-package me.digi.ongoingpostbox.features.base
-
-import androidx.appcompat.app.AppCompatActivity
-
-abstract class BaseActivity : AppCompatActivity()
