+++ examples/ongoingpostbox/src/main/java/me/digi/ongoingpostbox/features/create/view/CreatePostboxFragment.kt
@@ -2,24 +2,13 @@ package me.digi.ongoingpostbox.features.create.view
 
 import android.os.Bundle
 import android.view.View
-import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.Observer
 import kotlinx.android.synthetic.main.fragment_create_postbox.*
 import me.digi.ongoingpostbox.R
-import me.digi.ongoingpostbox.domain.OngoingPostboxResponseBody
-import me.digi.ongoingpostbox.features.create.viewmodel.CreatePostboxViewModel
 import me.digi.ongoingpostbox.features.upload.view.UploadContentFragment
-import me.digi.ongoingpostbox.utils.Resource
 import me.digi.ongoingpostbox.utils.replaceFragment
-import me.digi.ongoingpostbox.utils.snackBar
-import org.koin.android.viewmodel.ext.android.viewModel
 
 class CreatePostboxFragment : Fragment(R.layout.fragment_create_postbox), View.OnClickListener {
-
-    private val viewModel: CreatePostboxViewModel by viewModel()
-
-    private var firstExecution: Boolean = true
 
     companion object {
         fun newInstance() = CreatePostboxFragment()
@@ -28,7 +17,6 @@ class CreatePostboxFragment : Fragment(R.layout.fragment_create_postbox), View.O
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        subscribeToObservers()
         handleClickListeners()
     }
 
@@ -36,41 +24,11 @@ class CreatePostboxFragment : Fragment(R.layout.fragment_create_postbox), View.O
         btnCreatePostbox?.setOnClickListener(this)
     }
 
-    private fun subscribeToObservers() {
-        viewModel.createPostboxStatus.observe(
-            viewLifecycleOwner,
-            Observer { result: Resource<OngoingPostboxResponseBody> ->
-                when (result) {
-                    is Resource.Loading -> {
-                        crePosProgressbar?.isVisible = true
-                        btnCreatePostbox?.isEnabled = false
-                    }
-                    is Resource.Success -> {
-                        crePosProgressbar?.isVisible = false
-                        handlePostboxSuccess(result.data!!)
-                    }
-                    is Resource.Failure -> {
-                        crePosProgressbar?.isVisible = false
-                        btnCreatePostbox?.isEnabled = true
-                        snackBar(result.message ?: "Unknown")
-                    }
-                }
-            })
-    }
-
-    private fun handlePostboxSuccess(result: OngoingPostboxResponseBody) =
-        UploadContentFragment
-            .newInstance(result)
-            .replaceFragment(requireActivity().supportFragmentManager)
-
     override fun onClick(view: View?) {
         when (view?.id) {
-            R.id.btnCreatePostbox -> {
-                if (firstExecution) {
-                    viewModel.createPostbox(requireActivity())
-                    firstExecution = false
-                }
-            }
+            R.id.btnCreatePostbox -> UploadContentFragment
+                .newInstance()
+                .replaceFragment(requireActivity().supportFragmentManager)
         }
     }
 }
