+++ examples/ongoingpostbox/src/main/java/me/digi/ongoingpostbox/features/upload/view/UploadContentFragment.kt
@@ -1,7 +1,6 @@
 package me.digi.ongoingpostbox.features.upload.view
 
 import android.app.Activity
-import android.content.Context
 import android.content.Intent
 import android.net.Uri
 import android.os.Bundle
@@ -14,37 +13,24 @@ import coil.transform.CircleCropTransformation
 import com.github.dhaval2404.imagepicker.ImagePicker
 import kotlinx.android.synthetic.main.fragment_upload_content.*
 import me.digi.ongoingpostbox.R
+import me.digi.ongoingpostbox.data.localaccess.MainLocalDataAccess
 import me.digi.ongoingpostbox.domain.OngoingPostboxResponseBody
-import me.digi.ongoingpostbox.features.upload.viewmodel.UploadContentViewModel
+import me.digi.ongoingpostbox.features.viewmodel.MainViewModel
 import me.digi.ongoingpostbox.utils.*
 import me.digi.sdk.entities.DMEMimeType
-import me.digi.sdk.entities.DMEOAuthToken
-import me.digi.sdk.entities.DMEPostbox
 import me.digi.sdk.entities.DMEPushPayload
+import org.koin.android.ext.android.inject
 import org.koin.android.viewmodel.ext.android.viewModel
-import timber.log.Timber
 import java.io.File
 
 class UploadContentFragment : Fragment(R.layout.fragment_upload_content), View.OnClickListener {
 
-    private val viewModel: UploadContentViewModel by viewModel()
-    private lateinit var ongoingResult: OngoingPostboxResponseBody
+    private val viewModel: MainViewModel by viewModel()
+    private val localAccess: MainLocalDataAccess by inject()
+    private var firstExecution = true
 
     companion object {
         fun newInstance() = UploadContentFragment()
-        fun newInstance(ongoingResult: OngoingPostboxResponseBody) = UploadContentFragment().apply {
-            arguments = Bundle().apply {
-                putParcelable("ongoingResult", ongoingResult)
-            }
-        }
-    }
-
-    override fun onAttach(context: Context) {
-        super.onAttach(context)
-
-        arguments?.getParcelable<OngoingPostboxResponseBody>("ongoingResult")?.let {
-            ongoingResult = it
-        }
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -60,6 +46,27 @@ class UploadContentFragment : Fragment(R.layout.fragment_upload_content), View.O
     }
 
     private fun subscribeToObservers() {
+        viewModel.createPostboxStatus.observe(
+            viewLifecycleOwner,
+            Observer { result: Resource<OngoingPostboxResponseBody> ->
+                when (result) {
+                    is Resource.Loading -> {
+                        pbUploadContent?.isVisible = true
+                        btnUploadImage?.isEnabled = false
+                    }
+                    is Resource.Success -> {
+                        pbUploadContent?.isVisible = false
+                        btnUploadImage?.isEnabled = true
+                        snackBar("Workspace ready")
+                    }
+                    is Resource.Failure -> {
+                        pbUploadContent?.isVisible = false
+                        btnUploadImage?.isEnabled = true
+                        snackBar(result.message ?: "Unknown")
+                    }
+                }
+            })
+
         viewModel.uploadDataStatus.observe(
             viewLifecycleOwner,
             Observer { result: Resource<Boolean> ->
@@ -72,7 +79,6 @@ class UploadContentFragment : Fragment(R.layout.fragment_upload_content), View.O
                         pbUploadContent?.isVisible = false
                         btnUploadImage?.isEnabled = true
                         snackBar("Image uploaded successfully")
-                        Timber.d("Image upload successful: ${result.data}")
                     }
                     is Resource.Failure -> {
                         pbUploadContent?.isVisible = false
@@ -86,7 +92,6 @@ class UploadContentFragment : Fragment(R.layout.fragment_upload_content), View.O
     private fun handleImageRequest(resultCode: Int, data: Intent?) {
         when (resultCode) {
             Activity.RESULT_OK -> {
-                //Image Uri will not be null for RESULT_OK
                 val fileUri: Uri = data?.data!!
 
                 ivImageToUpload?.load(fileUri) {
@@ -95,24 +100,22 @@ class UploadContentFragment : Fragment(R.layout.fragment_upload_content), View.O
 
                 //You can get File object from intent
                 val file: File = ImagePicker.getFile(data)!!
+
+                // Update UI based on the file existence
                 tvImageName?.text = file.nameWithoutExtension
                 tvMimeType?.isVisible = true
                 tvMimeType?.text = getMimeType(requireContext(), fileUri).toString()
                 btnUploadImage?.isEnabled = true
 
+                /**
+                 * At this point we have both
+                 * @see [localAccess.getCachedPostbox]
+                 * @see [localAccess.getCachedCredential]
+                 * so we can say they're not null
+                 */
                 val metadata = getFileContent(requireActivity(), "metadatapng.json")
-                val postbox = DMEPostbox(
-                    ongoingResult.sessionKey!!,
-                    ongoingResult.postboxId!!,
-                    ongoingResult.publicKey!!,
-                    ongoingResult.digiMeVersion
-                )
-                val credentials = DMEOAuthToken(
-                    ongoingResult.accessToken!!,
-                    ongoingResult.expiresOn!!,
-                    ongoingResult.refreshToken!!,
-                    ongoingResult.tokenType!!
-                )
+                val postbox = localAccess.getCachedPostbox()!!
+                val credentials = localAccess.getCachedCredential()!!
                 val postboxPayload = DMEPushPayload(
                     postbox,
                     metadata,
@@ -136,6 +139,14 @@ class UploadContentFragment : Fragment(R.layout.fragment_upload_content), View.O
                 .compress(1024)
                 .crop(16f, 9f)
                 .start(::handleImageRequest)
+        }
+    }
+
+    override fun onResume() {
+        super.onResume()
+        if (firstExecution) {
+            viewModel.createPostbox(requireActivity())
+            firstExecution = false
         }
     }
 }
