+++ examples/ongoingpostbox/src/main/res/drawable/ic_add.xml
@@ -1,10 +1,13 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="24"
-    android:viewportHeight="24"
-    android:tint="?attr/colorControlNormal">
+    android:width="29dp"
+    android:height="29dp"
+    android:viewportWidth="29"
+    android:viewportHeight="29">
   <path
-      android:fillColor="@android:color/white"
-      android:pathData="M19,13h-6v6h-2v-6H5v-2h6V5h2v6h6v2z"/>
+      android:strokeWidth="1"
+      android:pathData="M14.5989,7.2008V21.2008M7.5989,14.2008H21.5989"
+      android:strokeLineJoin="round"
+      android:fillColor="#00000000"
+      android:strokeColor="#ecf0f1"
+      android:strokeLineCap="round"/>
 </vector>
