+++ examples/ongoingpostbox/src/main/res/layout/fragment_upload_content.xml
@@ -25,6 +25,7 @@
         android:layout_marginEnd="16dp"
         android:layout_marginBottom="16dp"
         android:clickable="true"
+        android:tint="@color/white"
         android:enabled="false"
         android:focusable="true"
         app:layout_constraintBottom_toBottomOf="parent"
@@ -44,10 +45,11 @@
 
         <ImageView
             android:id="@+id/ivImageToUpload"
-            android:layout_width="150dp"
-            android:layout_height="150dp"
+            android:layout_width="100dp"
+            android:layout_height="100dp"
             android:layout_gravity="center"
-            tools:srcCompat="@tools:sample/backgrounds/scenic" />
+            app:srcCompat="@drawable/ic_upload"
+            tools:srcCompat="@drawable/ic_upload" />
 
         <TextView
             android:id="@+id/tvImageName"
@@ -57,7 +59,7 @@
             android:layout_marginBottom="8dp"
             android:gravity="center"
             android:text="Tap on image to select one"
-            android:textSize="18sp"
+            android:textSize="16sp"
             tools:text="Image name" />
 
         <TextView
