+++ examples/ongoingpostbox/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">ongoingpostbox</string>
+    <string name="app_name">CPU</string>
 
     <string name="digime_application_id">IfnN9Y27Jym3P1Fad3ks3sTlo22flUBb</string>
     <string name="digime_contract_id">V5cRNEhdXHWqDEM54tZNqBaElDQcfl4v</string>
